requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
supabase>=2.10.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.27.2